requirements.txt
Scrapy==1.4.0
scrapy-s3pipeline==0.1.1
